Reorganize options package
